# docs/openapi.yaml
components:
  schemas:
    AddUserRoleBody:
      description: add user request body
      properties:
        role_id:
          title: Role Id
          type: string
        user_id:
          title: User Id
          type: string
      required:
      - role_id
      - user_id
      title: AddUserRoleBody
      type: object
...
    ClientScopesBody:
      description: Add client scopes body
      properties:
        client_id:
          format: uuid
          title: Client Id
          type: string
        roles:
          items:
            type: string
          title: Roles
          type: array
      required:
      - client_id
      - roles
      title: ClientScopesBody
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RoleAddPatchBody:
      description: Initial role add body
      properties:
        id_:
          title: 'Id '
          type: string
        scopes:
          items:
            type: string
          title: Scopes
          type: array
      required:
      - id_
      - scopes
      title: RoleAddPatchBody
      type: object
    RoleScopesBody:
      description: Add user scope request body
      properties:
        role_id:
          title: Role Id
          type: string
        scope_id:
          title: Scope Id
          type: string
      required:
      - role_id
      - scope_id
      title: RoleScopesBody
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: Jwks Authentication API
  title: Auth Api
  version: '1.0'
openapi: 3.1.0
paths:
...
  /clients/client/role:
    delete:
      description: delete client scopes
      operationId: delete_client_roles_clients_client_role_delete
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientScopesBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Client Roles
      tags:
      - Clients Authenticated
    post:
      description: add client roles
      operationId: add_client_role_clients_client_role_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientScopesBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Client Role
      tags:
      - Clients Authenticated
  /clients/client/roles:
    get:
      description: get client roles
      operationId: get_client_roles_clients_client_roles_get
      parameters:
      - in: query
        name: client_id
        required: true
        schema:
          format: uuid
          title: Client Id
          type: string
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Client Roles
      tags:
      - Clients Authenticated
...
  /roles:
    get:
      description: get roles
      operationId: get_roles_roles_get
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Roles
      tags:
      - Roles Authenticated
  /roles/role:
    delete:
      description: delete roles
      operationId: delete_role_roles_role_delete
      parameters:
      - in: query
        name: id_
        required: true
        schema:
          title: 'Id '
          type: string
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Role
      tags:
      - Roles Authenticated
    patch:
      description: Add role with scopes
      operationId: update_role_roles_role_patch
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAddPatchBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Role
      tags:
      - Roles Authenticated
    post:
      description: Add role with scopes
      operationId: add_role_roles_role_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAddPatchBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Role
      tags:
      - Roles Authenticated
  /roles/role/scopes:
    delete:
      description: get user scopes
      operationId: delete_role_scope_roles_role_scopes_delete
      parameters:
      - in: query
        name: role_id
        required: true
        schema:
          title: Role Id
          type: string
      - in: query
        name: scope_id
        required: true
        schema:
          title: Scope Id
          type: string
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Role Scope
      tags:
      - Roles Authenticated
    get:
      description: get user scopes
      operationId: get_role_scopes_roles_role_scopes_get
      parameters:
      - in: query
        name: role_id
        required: true
        schema:
          title: Role Id
          type: string
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Role Scopes
      tags:
      - Roles Authenticated
    post:
      description: add role scopes
      operationId: add_role_scopes_roles_role_scopes_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleScopesBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Role Scopes
      tags:
      - Roles Authenticated
...
  /users/user/roles:
    post:
      description: create new user endpoint. stores a hash of the password
      operationId: add_user_role_users_user_roles_post
      parameters:
      - in: header
        name: user-agent
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User-Agent
      - in: cookie
        name: session_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserRoleBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add User Role
      tags:
      - Users Authenticated
...